Improve failing test feedback
A number of our tests result in asserting equality with a null object
when the test fails. While this works, it doesn't provide any
information as to the cause of the test failure such as an exception.
Additionally, the way those tests are written is more verbose than
necessary, making them hard to follow. This commit improves some
low-hanging fruit test by making them less verbose and more informative
when they fail.