Implement app download metrics collection
As part of this change, we've created our own ClickHouse Quarkus
extensions to:

- Improve our developer experience with a ClickHouse Quarkus Dev Service
- Unify data source configuration
- Support GraalVM Native Image compilation

The existing Quarkiverse extension is unmaintained, years out of date,
and broken, making it unviable for us to use instead.

quarkus-jdbc-clickhouse includes a brotli decoding dependency. This may
appear unnecessary, but is actually required for native image
compilation to succeed due to clickhouse-jdbc's transitive dependency on
the Apache HTTP client. See
https://github.com/apache/camel-quarkus/issues/4512 for more information
on a similar issue.

Closes #16