Fix bug causing incorrect GetAppListing() responses
When calling GetAppListing() in such a way that the app's default
language listing is fallen back to, the listing information for the
first app ID in the database is returned instead of the requested app ID
(except the app's version information, which is returned correctly).

This bug occurs because Listing.findDefaultForApp() contains an
erroneous HQL query which returns a bogus list of listings from which
findDefaultForApp() selects the first entry since it only expects a
single entry. Release channel information (such as the app's version) is
not affected since a separate query is used to retrieve it.

This change fixes the bug by correcting the HQL query to return the app
listing associated with the app's default listing language. A test was
also added to ensure this behavior doesn't occur again.