Split & rename internal proto package
Using "internal" in our package name prevents Go clients from consuming
it as per
https://buf.build/docs/best-practices/style-guide/#recommendations.
Rename the "internal" package to "push" to resolve this issue and to
indicate that this package includes a separate service for pushing
updates to the app directory. Additionally split out
ListAppListingsPageToken into a separate "priv" package to properly
indicate that both 1) it should be opaque to clients and 2) it is not
related to the push directory service.