Accept new apps from Kafka instead of gRPC
Having a separate gRPC API for publishing new apps to the directory
service has a number of complexity issues, namely:

1. Quarkus cannot run the "push" API on a different socket than the
   directory service, so we need to either 1) rely on L7 network
   filtering to prevent external clients from accessing this privileged
   API or 2) implement a separate authn/authz system to ensure only
   appropriate internal services can access the API.
2. It very tightly couples Parcelo to publishing apps to the directory
   service, whereas we may want to be performing other actions based on
   this API separately, e.g., generating delta updates and/or logging
   metrics.

We are more broadly reorienting the directory service (and the
Accrescent server architecture as a whole) to an event-driven
architecture, so Kafka makes sense for the backbone of this system. We
can run in in our private network and apply Kafka ACLs to control
access, solving (1), and since it'll be the backbone of our architecture
moving forward, we don't need to worry about (2) (i.e. we'll be coupled
to Kafka, but that's okay).