Add date to events via constructor bodies
Adding the date to events automatically via their respective
constructor bodies instead of manually as constructor parameters ensures
we can't erroneously set the date to some past or future date. It also
ensures we also use UTC for the date timezone, which is important for
ensuring our event dating is consistent across all events so that we can
correlate them with each other.