Migrate server events to protovalidate validation
Because validation is delegated to protovalidate and the server-events
protobuf schema, I think it's reasonable to limit our validation tests
in the directory server to only what it is responsible for, that is,
reporting errors properly on an invalid message. Testing every invalid
message configuration isn't necessary here since that responsibility
belongs to the server-events schema and protovalidate.