Use Kafka EOS for publication request processing
This change ensures our request processing is safe and atomic.

It is more difficult for us to implement dead letter queues using EOS in
this way, so dead letter queue support is removed for now. A
deserialization error is indicative of a serious bug in an Accrescent
server component, so we do not assume for now that message processing
can continue after sending the offending message to the dead letter
queue. We probably want to change this in the future for availability
reasons in case we _do_ ship a serious bug, but for now, we'll do the
easy and obvious thing by causing a fatal error if we can't deserialize
an incoming request from another server component.