Remove ClickHouse Quarkus extension
Since #45 ("Send events to message broker instead of ClickHouse
directly") and fa5cfcc ("Remove unused metric events"), we no longer
need ClickHouse in our architecture. We may need it in the future, but
for now, we have no need to maintain a Quarkus extension for ClickHouse,
and it is effectively unmaintained since we're not actively using it.

We will likely revive the extension in the future if we have need of it.
But for now, it makes sense for us to remove it.

Closes #44