Fix native build crashing on startup
protovalidate-java uses reflection, but has no direct support for
GraalVM or plans to implement such (see
bufbuild/protovalidate-java#206), which causes the directory server to
crash on startup.

Add the necessary reflection config so that native compilation with
GraalVM works as expected.

This change was tested manually by running the native image build of the
directory server using this repository's Dockerfile. Ideally we would
test this automatically in CI with native Quarkus integration tests, but
that would be significantly more work to set up, so we're just fixing
the startup crash issue for now.