Remove unnecessary internal checks in download methods
Checking if the APK list returned from the database is empty does check
a valid error case. However, the check has two issues: it is not
comprehensive, and it is unnecessarily checking for a serious internal
error.

The check is not comprehensive because it is still possible for the APK
list returned from the database to contain elements and still be wrong.
We _could_ expand the check to ensure it covers this case too; however,
that would involve distrusting our database, which we must rely on for
proper operation. In other words, if we check for invalid data here, why
not check it for _every_ database query? That, obviously, doesn't make
sense, so this change removes these unnecessary checks to simplify our
code.